// src/features/doctors/types.ts
export interface Doctor {
  id: string
  name: string
  specialty: string
  consultationDuration: number
  workingDays: number[]
  startTime: string
  endTime: string
}
// export interface Doctor {
//   id: string
//   name: string
//   specialty: string
//   consultationDuration: number
//   workStart: string
//   workEnd: string
// }